+++ src/@types/custom-types.ts
@@ -1,0 +1,1 @@
+export type GeoCoordinates = [string | number, string | number]; //latitude & longitude