+++ src/@types/interfaces.ts
@@ -1,0 +1,17 @@
+import { UnitType } from './enum';
+
+export interface KeyValuePair {
+  [key: string]: any;
+}
+
+export interface Availability {
+  start?: string | number;
+  end?: string | number;
+}
+
+export interface TimelinesOptions {
+  fields: string[];
+  availability?: Availability;
+  units?: UnitType;
+  timezone?: string;
+}