+++ src/browser.ts
@@ -1,7 +1,23 @@
-import { Main } from "./main";
+/**
+ * * Classes
+ */
+import { Main } from './main';
+
+/**
+ * * Types
+ */
+import { GeoCoordinates } from './@types';
+import { QueryBuilderOptions } from './util/@types/interface';
 
 export default class ClimaCellAPI extends Main {
-  constructor() {
-    super(window.fetch.bind(window));
+  /**
+   * * Static fields
+   */
+  static async requestData(options: QueryBuilderOptions): Promise<any> {
+    return super.requestData(window.fetch.bind(window), options);
+  }
+
+  constructor(apiKey: string, coordinates: GeoCoordinates) {
+    super(window.fetch.bind(window), apiKey, coordinates);
   }
 }