+++ src/node.ts
@@ -1,7 +1,23 @@
+/**
+ * * Classes
+ */
 import { Main } from './main';
 
+/**
+ * * Types
+ */
+import { GeoCoordinates } from './@types';
+import { QueryBuilderOptions } from './util/@types/interface';
+
 export default class ClimaCellAPI extends Main {
-  constructor() {
-    super(require('node-fetch'));
+  /**
+   * * Static fields
+   */
+  static async requestData(options: QueryBuilderOptions): Promise<any> {
+    return super.requestData(require('node-fetch'), options);
+  }
+
+  constructor(apiKey: string, coordinates: GeoCoordinates) {
+    super(require('node-fetch'), apiKey, coordinates);
   }
 }