+++ webpack.config.js
@@ -34,9 +34,7 @@ const nodeConfig = {
   output: {
     path: path.resolve(__dirname, './dist'),
     filename: 'node.js',
-    library: 'ClimaCellAPI',
     libraryTarget: 'umd',
-    umdNamedDefine: true,
     libraryExport: 'default',
   },
 };
@@ -47,11 +45,11 @@ const browserConfig = {
   output: {
     path: path.resolve(__dirname, './dist'),
     filename: 'browser.js',
-    library: 'ClimaCellAPI',
     libraryTarget: 'umd',
     globalObject: 'this',
-    umdNamedDefine: true,
     libraryExport: 'default',
+    umdNamedDefine: true,
+    library: 'ClimaCellAPI',
   },
 };